+++ src/graphics/attributeSize.mjs
