+++ src/graphics/attributeType.mjs
