+++ src/polygonData.mjs
@@ -1,4 +1,4 @@
-import AttributeTypes from "./graphics/attributeTypes.mjs";
+import AttributeType from "./graphics/attributeType.mjs";
 import AttributeSchema from "./graphics/attributeSchema.mjs";
 import AttributeElement from "./graphics/attributeElement.mjs";
 import VertexUsage from "./graphics/vertexUsage.mjs";
@@ -6,7 +6,7 @@ import VertexBuffer from "./graphics/vertexBuffer.mjs";
 import IndexBuffer from "./graphics/indexBuffer.mjs";
 
 const attributeSchema = new AttributeSchema([
-    new AttributeElement("a_position", 3, AttributeTypes.FLOAT)
+    new AttributeElement("a_position", 3, AttributeType.FLOAT)
 ]);
 
 export default class PolygonData {
